+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Changed
+
+ - Properly extract test messages and report test failure errors
+
 # 0.2.0 - 2024-12-16
 
 ## Added